api/requirements.txt
fastapi==0.101.0
uvicorn
python-dotenv~=1.0.0
typer==0.9.0
python-multipart
tortoise-orm[asyncpg]==0.19.3
pydantic==1.10.12
gunicorn==21.2.0
starlette~=0.27.0
aiofiles~=23.2.1
arrow~=1.2.3
aio-pika
loguru